#245 Rename 'transient' -> 'transient_eager' and 'on_demand' -> 'transient' since 'transient_eager' is going to be deprecated.
