+++ pramen/api/src/main/scala/za/co/absa/pramen/api/DataFormat.scala
@@ -21,46 +21,60 @@ sealed trait DataFormat {
   def name: String
 
   def isTransient: Boolean
+
+  def isLazy: Boolean
 }
 
 object DataFormat {
   case class Parquet(path: String, recordsPerPartition: Option[Long]) extends DataFormat {
     override def name: String = "parquet"
 
-    val isTransient: Boolean = false
+    override val isTransient: Boolean = false
+
+    override val isLazy: Boolean = false
   }
 
   case class Delta(query: Query, recordsPerPartition: Option[Long]) extends DataFormat {
     override def name: String = "delta"
 
-    val isTransient: Boolean = false
+    override val isTransient: Boolean = false
+
+    override val isLazy: Boolean = false
   }
 
   // This format is used for metatables which are just files and can only be used for further sourcing
   case class Raw(path: String) extends DataFormat {
     override def name: String = "raw"
 
-    val isTransient: Boolean = false
+    override val isTransient: Boolean = false
+
+    override val isLazy: Boolean = false
   }
 
   // This format is used for tables that exist only for the duration of the process, and is not persisted
-  case class Transient(cachePolicy: CachePolicy) extends DataFormat {
-    override def name: String = "transient"
+  case class TransientEager(cachePolicy: CachePolicy) extends DataFormat {
+    override def name: String = "transient_eager"
 
-    val isTransient: Boolean = true
+    override val isTransient: Boolean = true
+
+    override val isLazy: Boolean = false
   }
 
   // This format is used for tables are calculated only if requested, and is not persisted
-  case class OnDemand(cachePolicy: CachePolicy) extends DataFormat {
-    override def name: String = "on_demand"
+  case class Transient(cachePolicy: CachePolicy) extends DataFormat {
+    override def name: String = "transient"
 
-    val isTransient: Boolean = true
+    override val isTransient: Boolean = true
+
+    override val isLazy: Boolean = true
   }
 
   // This format is used for metatables which do not support persistence, e.g. for sink or transfer jobs
   case class Null() extends DataFormat {
     override def name: String = "null"
 
-    val isTransient: Boolean = false
+    override val isTransient: Boolean = false
+
+    override val isLazy: Boolean = false
   }
 }
