+++ pramen/core/src/test/resources/test/config/integration_transient_eager_transformer.conf
@@ -1,0 +1,118 @@
+# Copyright 2022 ABSA Group Limited
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+# This variable is expected to be set up by the test suite
+#base.path = "/tmp"
+
+pramen {
+  pipeline.name = "Integration test with a file-based source"
+
+  temporary.directory = ${base.path}/temp
+
+  bookkeeping.enabled = false
+  stop.spark.session = false
+}
+
+pramen.metastore {
+  tables = [
+    {
+      name = "table1"
+      description = "Table 1 (file based)"
+      format = "raw"
+      path = ${base.path}/table1
+    },
+    {
+      name = "table2"
+      description = "Table 2 (transient)"
+      format = "transient_eager"
+      cache.policy = ${cache.policy}
+    }
+    {
+      name = "table3"
+      description = "Table 3 (parquet)"
+      format = "parquet"
+      path = ${base.path}/table3
+    }
+  ]
+}
+
+pramen.sources.1 = [
+  {
+    name = "file_source"
+    factory.class = "za.co.absa.pramen.core.source.RawFileSource"
+  }
+]
+
+pramen.operations = [
+  {
+    name = "Sourcing from a folder"
+    type = "ingestion"
+    schedule.type = "daily"
+
+    source = "file_source"
+
+    info.date.expr = "@runDate"
+
+    tables = [
+      {
+        input.path = ${base.path}
+        output.metastore.table = table1
+      }
+    ]
+  },
+  {
+    name = "Converting to dataframe"
+    type = "transformation"
+
+    class = "za.co.absa.pramen.core.transformers.ConversionTransformer"
+    schedule.type = "daily"
+
+    output.table = "table2"
+
+    dependencies = [
+      {
+        tables = [ table1 ]
+        date.from = "@infoDate"
+        optional = true # Since no bookkeeping available the table will be seen as empty for the dependency manager
+      }
+    ]
+
+    option {
+      input.table = "table1"
+      input.format = "csv"
+
+      header = true
+    }
+  },
+  {
+    name = "Identity transformer"
+    type = "transformation"
+
+    class = "za.co.absa.pramen.core.transformers.IdentityTransformer"
+    schedule.type = "daily"
+
+    output.table = "table3"
+
+    dependencies = [
+      {
+        tables = [ table2 ]
+        date.from = "@infoDate"
+        optional = true # Since no bookkeeping available the table will be seen as empty for the dependency manager
+      }
+    ]
+
+    option {
+      table = "table2"
+    }
+  }
+]
