+++ pramen/core/src/test/resources/test/config/integration_transient_transformer.conf
@@ -27,57 +27,53 @@ pramen.metastore {
   tables = [
     {
       name = "table1"
-      description = "Table 1 (file based)"
-      format = "raw"
+      format = "parquet"
       path = ${base.path}/table1
     },
     {
       name = "table2"
-      description = "Table 2 (transient)"
       format = "transient"
       cache.policy = ${cache.policy}
-    }
+    },
     {
       name = "table3"
-      description = "Table 3 (parquet)"
       format = "parquet"
       path = ${base.path}/table3
+    },
+    {
+      name = "table4"
+      format = "parquet"
+      path = ${base.path}/table4
+    },
+    {
+      name = "table5"
+      format = "parquet"
+      path = ${base.path}/table5
     }
   ]
 }
 
-pramen.sources.1 = [
-  {
-    name = "file_source"
-    factory.class = "za.co.absa.pramen.core.source.RawFileSource"
-  }
-]
-
 pramen.operations = [
   {
-    name = "Sourcing from a folder"
-    type = "ingestion"
-    schedule.type = "daily"
-
-    source = "file_source"
-
-    info.date.expr = "@runDate"
-
-    tables = [
-      {
-        input.path = ${base.path}
-        output.metastore.table = table1
-      }
-    ]
-  },
-  {
-    name = "Converting to dataframe"
+    name = "Generating dataframe"
     type = "transformation"
 
-    class = "za.co.absa.pramen.core.transformers.ConversionTransformer"
+    class = "za.co.absa.pramen.core.mocks.transformer.GeneratingTransformer"
     schedule.type = "daily"
+    output.table = "table1"
+    info.date.expr = "@runDate - 1"
+  },
+  {
+    name = "Identity transformer (on-demand)"
+    type = "transformation"
+
+    class = "za.co.absa.pramen.core.transformers.IdentityTransformer"
+    schedule.type = "weekly"
+    schedule.days.of.week = [ 1, 4 ] # Monday, Thursday
 
     output.table = "table2"
+
+    info.date.expr = "@runDate - 1"
 
     dependencies = [
       {
@@ -88,21 +84,18 @@ pramen.operations = [
     ]
 
     option {
-      input.table = "table1"
-      input.format = "csv"
-
-      header = true
+      table = "table1"
     }
   },
   {
-    name = "Identity transformer"
+    name = "Identity transformer (eager)"
     type = "transformation"
-
     class = "za.co.absa.pramen.core.transformers.IdentityTransformer"
     schedule.type = "daily"
 
     output.table = "table3"
 
+    info.date.expr = "@runDate - 1"
     dependencies = [
       {
         tables = [ table2 ]
@@ -113,6 +106,56 @@ pramen.operations = [
 
     option {
       table = "table2"
+    }
+  },
+  {
+    name = "Ranged transformer"
+    type = "transformation"
+
+    class = "za.co.absa.pramen.core.mocks.transformer.RangedTransformer"
+    schedule.type = "weekly"
+    schedule.days.of.week = [ 4 ] # Thursday
+
+    output.table = "table4"
+
+    info.date.expr = "@runDate"
+
+    dependencies = [
+      {
+        tables = [ table2 ]
+        date.from = "@infoDate - 1"
+        optional = true # Since no bookkeeping available the table will be seen as empty for the dependency manager
+      }
+    ]
+
+    option {
+      table = "table2"
+      use.latest = "false"
+    }
+  },
+  {
+    name = "Last snapshot transformer"
+    type = "transformation"
+
+    class = "za.co.absa.pramen.core.mocks.transformer.RangedTransformer"
+    schedule.type = "weekly"
+    schedule.days.of.week = [ 4 ] # Thursday
+
+    output.table = "table5"
+
+    info.date.expr = "@runDate"
+
+    dependencies = [
+      {
+        tables = [ table2 ]
+        date.from = "@infoDate - 1"
+        optional = true # Since no bookkeeping available the table will be seen as empty for the dependency manager
+      }
+    ]
+
+    option {
+      table = "table2"
+      use.latest = "true"
     }
   }
 ]
