+++ pramen/core/src/test/scala/za/co/absa/pramen/core/metastore/persistence/MetastorePersistenceTransientEagerSuiteEager.scala
@@ -22,11 +22,11 @@ import org.scalatest.wordspec.AnyWordSpec
 import za.co.absa.pramen.api.CachePolicy
 import za.co.absa.pramen.core.base.SparkTestBase
 import za.co.absa.pramen.core.fixtures.TempDirFixture
-import za.co.absa.pramen.core.metastore.peristence.MetastorePersistenceTransient
+import za.co.absa.pramen.core.metastore.peristence.MetastorePersistenceTransientEager
 
 import java.time.LocalDate
 
-class MetastorePersistenceTransientSuite extends AnyWordSpec with BeforeAndAfterAll with SparkTestBase with TempDirFixture {
+class MetastorePersistenceTransientEagerSuiteEager extends AnyWordSpec with BeforeAndAfterAll with SparkTestBase with TempDirFixture {
 
   import spark.implicits._
 
@@ -49,12 +49,12 @@ class MetastorePersistenceTransientSuite extends AnyWordSpec with BeforeAndAfter
   }
 
   "reset should do nothing if spark session is not available" in {
-    MetastorePersistenceTransient.reset()
+    MetastorePersistenceTransientEager.reset()
   }
 
   "loadTable" should {
     "return data for a date when it is available" in {
-      val persistor = new MetastorePersistenceTransient(null, "table1", CachePolicy.NoCache)
+      val persistor = new MetastorePersistenceTransientEager(None, "table1", CachePolicy.NoCache)
 
       persistor.saveTable(infoDate, exampleDf, Some(10))
 
@@ -62,11 +62,11 @@ class MetastorePersistenceTransientSuite extends AnyWordSpec with BeforeAndAfter
 
       assert(df.count() == 3)
 
-      MetastorePersistenceTransient.reset()
+      MetastorePersistenceTransientEager.reset()
     }
 
     "return an empty dataframe if data is not available but schema is available" in {
-      val persistor = new MetastorePersistenceTransient(null, "table2", CachePolicy.Cache)
+      val persistor = new MetastorePersistenceTransientEager(None, "table2", CachePolicy.Cache)
 
       persistor.saveTable(infoDate, exampleDf, Some(10))
 
@@ -74,11 +74,11 @@ class MetastorePersistenceTransientSuite extends AnyWordSpec with BeforeAndAfter
 
       assert(df.isEmpty)
 
-      MetastorePersistenceTransient.reset()
+      MetastorePersistenceTransientEager.reset()
     }
 
     "throw an exception if the data and schema are not available" in {
-      val persistor = new MetastorePersistenceTransient(null, "table2", CachePolicy.Cache)
+      val persistor = new MetastorePersistenceTransientEager(None, "table2", CachePolicy.Cache)
 
       val ex = intercept[IllegalStateException] {
         persistor.loadTable(Some(infoDate.plusDays(1)), Some(infoDate.plusDays(1)))
@@ -86,11 +86,11 @@ class MetastorePersistenceTransientSuite extends AnyWordSpec with BeforeAndAfter
 
       assert(ex.getMessage.contains("No data for transient table 'table2' for '2022-02-19'"))
 
-      MetastorePersistenceTransient.reset()
+      MetastorePersistenceTransientEager.reset()
     }
 
     "throw an exception on range queries" in {
-      val persistor = new MetastorePersistenceTransient(null, "table2", CachePolicy.Cache)
+      val persistor = new MetastorePersistenceTransientEager(None, "table2", CachePolicy.Cache)
 
       persistor.saveTable(infoDate, exampleDf, None)
 
@@ -100,11 +100,11 @@ class MetastorePersistenceTransientSuite extends AnyWordSpec with BeforeAndAfter
 
       assert(ex.getMessage.contains("Metastore 'transient' format does not support ranged queries"))
 
-      MetastorePersistenceTransient.reset()
+      MetastorePersistenceTransientEager.reset()
     }
 
     "throw an exception if info date is not provided" in {
-      val persistor = new MetastorePersistenceTransient(null, "table2", CachePolicy.Cache)
+      val persistor = new MetastorePersistenceTransientEager(None, "table2", CachePolicy.Cache)
 
       val ex = intercept[IllegalArgumentException] {
         persistor.loadTable(None, None)
@@ -112,35 +112,35 @@ class MetastorePersistenceTransientSuite extends AnyWordSpec with BeforeAndAfter
 
       assert(ex.getMessage.contains("Metastore 'transient' format requires info date for querying its contents"))
 
-      MetastorePersistenceTransient.reset()
+      MetastorePersistenceTransientEager.reset()
     }
   }
 
   "saveTable" should {
     "work with non-cached data frames" in {
-      val persistor = new MetastorePersistenceTransient(null, "table_not_cached", CachePolicy.NoCache)
+      val persistor = new MetastorePersistenceTransientEager(None, "table_not_cached", CachePolicy.NoCache)
 
       val saveResult = persistor.saveTable(infoDate, exampleDf, Some(10))
 
       assert(saveResult.recordCount == 10)
       assert(saveResult.dataSizeBytes.isEmpty)
 
-      MetastorePersistenceTransient.reset()
+      MetastorePersistenceTransientEager.reset()
     }
 
     "work with cached data frames" in {
-      val persistor = new MetastorePersistenceTransient(null, "table_cached", CachePolicy.Cache)
+      val persistor = new MetastorePersistenceTransientEager(None, "table_cached", CachePolicy.Cache)
 
       val saveResult = persistor.saveTable(infoDate, exampleDf, None)
 
       assert(saveResult.recordCount == 3)
       assert(saveResult.dataSizeBytes.isEmpty)
 
-      MetastorePersistenceTransient.reset()
+      MetastorePersistenceTransientEager.reset()
     }
 
     "work with persisted data frames" in {
-      val persistor = new MetastorePersistenceTransient(tempDir, "table_persisted", CachePolicy.Persist)
+      val persistor = new MetastorePersistenceTransientEager(Some(tempDir), "table_persisted", CachePolicy.Persist)
 
       val saveResult = persistor.saveTable(infoDate, exampleDf, None)
 
@@ -148,13 +148,13 @@ class MetastorePersistenceTransientSuite extends AnyWordSpec with BeforeAndAfter
       assert(saveResult.dataSizeBytes.isDefined)
       assert(saveResult.dataSizeBytes.exists(_ > 100))
 
-      MetastorePersistenceTransient.reset()
+      MetastorePersistenceTransientEager.reset()
     }
   }
 
   "getStats" should {
     "not be supported" in {
-      val persistor = new MetastorePersistenceTransient(null, null, null)
+      val persistor = new MetastorePersistenceTransientEager(null, null, null)
 
       assertThrows[UnsupportedOperationException] {
         persistor.getStats(null)
@@ -164,7 +164,7 @@ class MetastorePersistenceTransientSuite extends AnyWordSpec with BeforeAndAfter
 
   "createOrUpdateHiveTable" should {
     "not be supported" in {
-      val persistor = new MetastorePersistenceTransient(null, null, null)
+      val persistor = new MetastorePersistenceTransientEager(null, null, null)
 
       assertThrows[UnsupportedOperationException] {
         persistor.createOrUpdateHiveTable(null, null, null, null)
@@ -174,7 +174,7 @@ class MetastorePersistenceTransientSuite extends AnyWordSpec with BeforeAndAfter
 
   "repairHiveTable" should {
     "not be supported" in {
-      val persistor = new MetastorePersistenceTransient(null, null, null)
+      val persistor = new MetastorePersistenceTransientEager(null, null, null)
 
       assertThrows[UnsupportedOperationException] {
         persistor.repairHiveTable(null, null, null)
@@ -184,107 +184,107 @@ class MetastorePersistenceTransientSuite extends AnyWordSpec with BeforeAndAfter
 
   "addRawDataFrame" should {
     "return dataframe only" in {
-      val (df, size) = MetastorePersistenceTransient.addRawDataFrame("table_not_cached2", infoDate, exampleDf)
+      val (df, size) = MetastorePersistenceTransientEager.addRawDataFrame("table_not_cached2", infoDate, exampleDf)
 
       assert(df.schema.treeString == exampleDf.schema.treeString)
       assert(df.count() == 3)
       assert(size.isEmpty)
 
-      MetastorePersistenceTransient.reset()
+      MetastorePersistenceTransientEager.reset()
     }
   }
 
   "cachedDataframes" should {
     "return dataframe only" in {
-      val (df, size) = MetastorePersistenceTransient.addCachedDataframe("table_cached2", infoDate, exampleDf)
+      val (df, size) = MetastorePersistenceTransientEager.addCachedDataframe("table_cached2", infoDate, exampleDf)
 
       assert(df.schema.treeString == exampleDf.schema.treeString)
       assert(df.count() == 3)
       assert(size.isEmpty)
 
-      MetastorePersistenceTransient.reset()
+      MetastorePersistenceTransientEager.reset()
     }
   }
 
   "addPersistedLocation" should {
     "return dataframe and size" in {
-      val (df, size) = MetastorePersistenceTransient.addPersistedDataFrame("table_persist2", infoDate, exampleDf, tempDir)
+      val (df, size) = MetastorePersistenceTransientEager.addPersistedDataFrame("table_persist2", infoDate, exampleDf, tempDir)
 
       assert(df.count() == 3)
       assert(size.isDefined)
       assert(size.exists(_ > 100))
 
-      MetastorePersistenceTransient.reset()
+      MetastorePersistenceTransientEager.reset()
     }
   }
 
   "getDataForTheDate" should {
     "work for non-cached dataframes" in {
-      MetastorePersistenceTransient.addRawDataFrame("table_not_cached3", infoDate, exampleDf)
+      MetastorePersistenceTransientEager.addRawDataFrame("table_not_cached3", infoDate, exampleDf)
 
-      assert(!MetastorePersistenceTransient.getDataForTheDate("table_not_cached3", infoDate).isEmpty)
+      assert(!MetastorePersistenceTransientEager.getDataForTheDate("table_not_cached3", infoDate).isEmpty)
 
-      MetastorePersistenceTransient.reset()
+      MetastorePersistenceTransientEager.reset()
     }
 
     "work for cached dataframes" in {
-      MetastorePersistenceTransient.addCachedDataframe("table_cached3", infoDate, exampleDf)
+      MetastorePersistenceTransientEager.addCachedDataframe("table_cached3", infoDate, exampleDf)
 
-      assert(!MetastorePersistenceTransient.getDataForTheDate("table_cached3", infoDate).isEmpty)
+      assert(!MetastorePersistenceTransientEager.getDataForTheDate("table_cached3", infoDate).isEmpty)
 
-      MetastorePersistenceTransient.reset()
+      MetastorePersistenceTransientEager.reset()
     }
 
     "work for persisted dataframes" in {
-      MetastorePersistenceTransient.addPersistedDataFrame("table_persist3", infoDate, exampleDf, tempDir)
+      MetastorePersistenceTransientEager.addPersistedDataFrame("table_persist3", infoDate, exampleDf, tempDir)
 
-      assert(!MetastorePersistenceTransient.getDataForTheDate("table_persist3", infoDate).isEmpty)
+      assert(!MetastorePersistenceTransientEager.getDataForTheDate("table_persist3", infoDate).isEmpty)
 
-      MetastorePersistenceTransient.reset()
+      MetastorePersistenceTransientEager.reset()
     }
 
     "return an empty dataframe if data not found but schema found" in {
-      MetastorePersistenceTransient.addCachedDataframe("table_cache", infoDate, exampleDf)
+      MetastorePersistenceTransientEager.addCachedDataframe("table_cache", infoDate, exampleDf)
 
-      val df = MetastorePersistenceTransient.getDataForTheDate("table_cache", infoDate.plusDays(1))
+      val df = MetastorePersistenceTransientEager.getDataForTheDate("table_cache", infoDate.plusDays(1))
 
       assert(df.isEmpty)
       assert(df.schema.sameElements(exampleDf.schema))
 
-      MetastorePersistenceTransient.reset()
+      MetastorePersistenceTransientEager.reset()
     }
 
     "throw an exception if data nor schema not found" in {
       assertThrows[IllegalStateException] {
-        MetastorePersistenceTransient.getDataForTheDate("table_cache", infoDate.plusDays(1))
+        MetastorePersistenceTransientEager.getDataForTheDate("table_cache", infoDate.plusDays(1))
       }
 
-      MetastorePersistenceTransient.reset()
+      MetastorePersistenceTransientEager.reset()
     }
   }
 
   "cleanup" should {
     "remove all types of transient tables from the internal state" in {
-      MetastorePersistenceTransient.addRawDataFrame("table_not_cached4", infoDate, exampleDf)
-      MetastorePersistenceTransient.addCachedDataframe("table_cached4", infoDate, exampleDf)
-      MetastorePersistenceTransient.addPersistedDataFrame("table_persist4", infoDate, exampleDf, tempDir)
+      MetastorePersistenceTransientEager.addRawDataFrame("table_not_cached4", infoDate, exampleDf)
+      MetastorePersistenceTransientEager.addCachedDataframe("table_cached4", infoDate, exampleDf)
+      MetastorePersistenceTransientEager.addPersistedDataFrame("table_persist4", infoDate, exampleDf, tempDir)
 
-      assert(!MetastorePersistenceTransient.getDataForTheDate("table_not_cached4", infoDate).isEmpty)
-      assert(!MetastorePersistenceTransient.getDataForTheDate("table_cached4", infoDate).isEmpty)
-      assert(!MetastorePersistenceTransient.getDataForTheDate("table_persist4", infoDate).isEmpty)
+      assert(!MetastorePersistenceTransientEager.getDataForTheDate("table_not_cached4", infoDate).isEmpty)
+      assert(!MetastorePersistenceTransientEager.getDataForTheDate("table_cached4", infoDate).isEmpty)
+      assert(!MetastorePersistenceTransientEager.getDataForTheDate("table_persist4", infoDate).isEmpty)
 
-      MetastorePersistenceTransient.reset()
+      MetastorePersistenceTransientEager.reset()
 
       assertThrows[IllegalStateException] {
-        MetastorePersistenceTransient.getDataForTheDate("table_not_cached4", infoDate)
+        MetastorePersistenceTransientEager.getDataForTheDate("table_not_cached4", infoDate)
       }
 
       assertThrows[IllegalStateException] {
-        MetastorePersistenceTransient.getDataForTheDate("table_cached4", infoDate)
+        MetastorePersistenceTransientEager.getDataForTheDate("table_cached4", infoDate)
       }
 
       assertThrows[IllegalStateException] {
-        MetastorePersistenceTransient.getDataForTheDate("table_persist4", infoDate)
+        MetastorePersistenceTransientEager.getDataForTheDate("table_persist4", infoDate)
       }
     }
   }
